# out/go/server/go/model_error.go
/*
 * Swagger Getting Started
 *
 * getting started
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

type Error struct {

	Code int32 `json:"code"`

	Message string `json:"message"`
}